Support Periodic Updates Of User Properties
Previously the addon allowed to update user properties only on every
request. For deployments in which the addon processes every request
to the site this option could not be used since the risk of
ConflicErrors was to high. This commit introduces a new option which
performs the update only in configurable intervals (one day by
default).